ref(performance): Make team key transaction dropdown pure component
This lifts the stateful parts of the team key transaction dropdown to the parent
component so it can be reused on else where.
